Don't depend on side effects of map() (#1141242)
In Python 2 calling map() returned a list, which had the
side effect of running the given function on all items in the
provided iterable, even if the result of the map() call was
not used.

On the other hand in Python 3 map() returns an iterator
and the mapped function is not applied on the iterable until
items are yielded from the iterator.

In short any map() usage that does not use the iterable
**does nothing** in Python 3.