Use a sane unified password checking policy
Use the same password checking policy in partial kickstart installation
as in interactive installation and clarify how the default password policy
is set and when it actually applies.

Also use a more sane policy - one that sets a minimal password
length but lets the user to override it.

Related: rhbz#1360263