Improved password quality checking
Refactor the password quality checking to make it more readable
and remove code duplication in the user and root password spokes.

The checks have been moved to the GUI helpers module and are used
in the user and root password spokes. In the future they can also be
used for luks passphrase checking and similar use cases.