Use structured logging for the anaconda logger
The logger naming aims to both make the logs easy to read and understand
while keeping the log prefixes reasonably stable over time.

For that reason the prefixes don't just always use the name of the file,
but use an easy to understand name when needed (eq. "custom_storage_config"
instead of just "custom" or "installation_source" instead of just "source").

The same thing is true for the logging hierarchy - we try to group
loggers for functional unit, not necessarily by folder. That's why all
non-hub & non-spoke modules live in ui.gui.* even though some are in
subfolders. In a similar manner the spoke prefix is ui.gui.spoke.* even
though the spokes are stored in a folder named "spokes".