Rename to User module to Users
Rename the module to Users to be more consistent with how the
other modules are named and to account for the fact it is not handling
more than a single user (not just root anymore).

This also has the additional benefit that we can have a submodule named
just "user" (for user configuration code) without creating a confusing folder
and module hierarchy.