Add system purpose support
The system purpose system is used to give hints to other tools about the
intended usage of the system.

There are two main interfaces:
- the syspurpose kickstart command
- the System Purpose spoke in GUI

Also as with other kickstart/spoke combos in Anaconda, the values
in kickstart should reflect in the GUI (if valid) and kickstart
can thus be used to pre-fill the GUI if needed.

As this functionality is not relevant for every distribution
it's usage is currently gated by an install class property.

Resolves: rhbz#1612060