Add support for SSH key management via the users DBUS module
Add support for reading in SSH key data from kickstart,
manipulating it via DBUS interface of the users DBUS module
and then writing it back into kickstart.

Also unit tests for all the above.

This should be the third and last kickstart command
(other than the already moved rootpw) that has to be
moved, so that we can move the Users object into a
DBUS module task for user/group creation.