subscription: Add service observer
Add ServiceObserver, which makes it possible to start
systemd unit backed DBus services on demand & to wait for
them to start up to a pre-defined timeout.

The timeout is set to 90 seconds by default, as this is also
the default timeout for systemd unit startup.

If the DBus service has not yet been requested, the corresponding
systemd unit will be started and get_proxy() will block up to
the timeout. If the DBus service becomes available before the timeout
DBus proxy for the service is cached and returned to the caller.

If the DBus service does not start in time a DBusObserverError is raised.

ServiceObserver is thread safe, so multiple threads can call
get_proxy() and safely wait for the DBus service to be available.

Once a proxy has been cached, subsequent get_proxy() calls should
not block ans will return the cached proxy.

On the other hand, if the service failed to start once, all subsequent
calls to get_proxy() will not block but will always raise the
DBusObserverError exception.