subscription: Add TransferSubscriptionTokensTask
Add the TransferSubscriptionTokensTask, which handles transfer
of subscription related tokens (basically various files) from
the installation environment to the target system.

This is needed due to the way how registration and subscription
is implemented in Anaconda - we first register and subscribe
the installation environment (so that it can talk to the CDN).
Due to this we need to transfer the respective subscription tokens
also to the target system or else it would not keep the attached
subscription after installation.

The subscription related files being transferred are:
- system purpose data json file
- the rhsm.conf configuration file
- the redhat.repo file describing what subscription repositories to use
- various key and certificate files that authenticate the system and its
  attached subscription to the subscription infrastructure and CDN