Fix CDN button visibility
Only show the Red Hat CDN button if the Subscription
module appears to be running. To achieve that, we do
the same thing as with the HMC button - the CDN
button invisible by default. And enable it only if it
looks like the Subscription module is running.

(cherry picked from commit: aca297587ab8bdd7342fa056273bfff379bc0a0b)

Resolves: rhbz#1903178