Initial Web UI support
This is achieved by creating a new UI type called
cockpit, for the time being living in parallel with
the existing GUI and TUI implementations.

But unlike the existing UIs, the cockpit UI will just
run the cockpit-desktop web view executable.

Also as we are adding a new module, we need to adjust our
autotools makefiles accordingly.

 ** the Web UI is currently optional **

To boot to the Web UI use the inst.webui boot option.

 ** cockpit-client VS cockpit-desktop **

Note that both cockpit-client and cockpit-desktop exist.
They both use the same webkit python browser code.

However, cockpit-desktop is available in the default install,
and cockpit-client is a flatpak.
We don't want to haul in all the flatpak stuff into boot.iso
and pre-install it when cockpit-desktop does the job just fine.

 ** UI run() method is expected to be blocking **

Both our existing UIs run their main loop in the run() method,
resulting in the method blocking as long as the UI main loop is running.

In our case we just start the web-view application in the run method,
so also add a very long sleep() call to emulate the expected run()
blocking behavior.