Build and install cockpit also through autotools
This commit includes cockpit files to the anaconda tarball instead of
generating a seperate one.

Now the NODE_ENV variable is always set to production in order to create
the expected files that the spec file expects.

(NODE_ENV production will compress the files in dist and do some
 minifications)

Wenn doing developement we want NODE_ENV=developement, which needs to
be supported in the future but I don't know how yet :)