Ignore npm packages files for translation
Without this the `all` target in `po/Makefile` will fail because there will be
too many files so shell complain about too long command line argument.

However, in general this may be a good place for improvement because right now
we are taking a lot more than we should as translation candidates.