HACK: drop brltty
At least in the AWS environment brltty can't find the /run/brltty/usbfs
device node and endlessly spams system log about it.

This pretty much prevents the cockpit-ws crash investigation we are
trying to do, as all relevant messages are quickly replaced by the
brltty spam.

So for now, drop the brltty dependency.